moved itemcreator to ItemStackUtils
